+++ theme-orange-country/build.gradle.kts
@@ -27,4 +27,7 @@ android {
 dependencies {
     implementation(project(":theme-contract"))
     implementation(project(":theme-orange"))
+
+    implementation(platform(libs.androidx.compose.bom))
+    implementation(libs.androidx.compose.material3)
 }
