--- theme-orange-derived/src/main/AndroidManifest.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-  ~ Software Name: OUDS Android
-  ~ SPDX-FileCopyrightText: Copyright (c) Orange SA
-  ~ SPDX-License-Identifier: MIT
-  ~
-  ~ This software is distributed under the MIT license,
-  ~ the text of which is available at https://opensource.org/license/MIT/
-  ~ or see the "LICENSE" file for more details.
-  ~
-  ~ Software description: Android library of reusable graphical components
-  -->
-
-<manifest />
